MeController - delete

+++ app/Controllers/Http/MeController.ts
@@ -1,4 +1,5 @@
 import type { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
+import APIException from 'App/Exceptions/APIException'
 import Log from 'App/Models/Log'
 
 export default class UsersController {
@@ -9,6 +10,12 @@ export default class UsersController {
 
   // Deletes the user.
   public async delete({ response, auth }: HttpContextContract) {
+    if (auth.user?.permission === 2) {
+      throw new APIException(
+        'Vous êtes un administrateur, votre compte ne peut pas être supprimé !',
+        403
+      )
+    }
     await auth.user!.delete()
     await auth.logout()
     return response.noContent()
@@ -16,16 +23,10 @@ export default class UsersController {
 
   public async update({ request, response, auth }: HttpContextContract) {
     // Retrieves the essential elements to update.
-    const { email, username, password } = request.only([
-      'email',
-      'username',
-      'password'
-    ])
+    const { email, username, password } = request.only(['email', 'username', 'password'])
 
     // Updates the user.
-    await auth.user!
-      .merge({ email, username, password })
-      .save()
+    await auth.user!.merge({ email, username, password }).save()
 
     return response.noContent()
   }
